Long type on losses

+++ nugraph/explain_graph/algorithms/linear_probes/accuracy_activation_vectors.py
@@ -47,7 +47,7 @@ class AccuracyActivationVectors(DynamicProbedNetwork):
         loss_func = RecallLoss(ignore_index=-1, num_classes=len(self.semantic_classes))
         labels = y.collect("y_semantic")
         for p in self.planes:
-            loss += loss_func(y_hat[p].type(torch.float), labels[p].type(torch.float))
+            loss += loss_func(y_hat[p].type(torch.long), labels[p].type(torch.long))
 
         return loss / len(self.planes)
 
@@ -55,10 +55,10 @@ class AccuracyActivationVectors(DynamicProbedNetwork):
         labels = y.collect("y_semantic")
         loss = 0
         for p in self.planes:
-            true_class = torch.tensor(labels[p] == class_index).type(torch.float)
+            true_class = torch.tensor(labels[p] == class_index).type(torch.long)
             prediction_class = torch.tensor(
                 torch.argmax(y_hat[p], axis=1) == class_index
-            ).type(torch.float)
+            ).type(torch.long)
             loss += torch.nn.functional.cross_entropy(prediction_class, true_class)
         return loss / len(self.planes)
